CRUD banner, coupon. Improve create order, delete some product, get product list for admin

+++ src/main/java/com/hcmute/drink/service/database/implement/ProductService.java
@@ -1,5 +1,6 @@
 package com.hcmute.drink.service.database.implement;
 
+import com.hcmute.drink.collection.CategoryCollection;
 import com.hcmute.drink.collection.ProductCollection;
 import com.hcmute.drink.collection.embedded.ImageEmbedded;
 import com.hcmute.drink.collection.embedded.SizeEmbedded;
@@ -10,16 +11,20 @@ import com.hcmute.drink.dto.request.UpdateProductRequest;
 import com.hcmute.drink.dto.response.*;
 import com.hcmute.drink.enums.ProductStatus;
 import com.hcmute.drink.model.CustomException;
+import com.hcmute.drink.model.elasticsearch.ProductIndex;
 import com.hcmute.drink.repository.database.ProductRepository;
 import com.hcmute.drink.service.database.IProductService;
 import com.hcmute.drink.service.elasticsearch.ProductSearchService;
-import com.hcmute.drink.utils.CloudinaryUtils;
+import com.hcmute.drink.service.common.CloudinaryService;
 import com.hcmute.drink.utils.ImageUtils;
-import com.hcmute.drink.utils.ModelMapperUtils;
+import com.hcmute.drink.service.common.ModelMapperService;
+import com.hcmute.drink.utils.RegexUtils;
+import com.hcmute.drink.utils.StringUtils;
 import lombok.RequiredArgsConstructor;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
+import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -27,16 +32,15 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
 import java.util.*;
 
-import static com.hcmute.drink.constant.ErrorConstant.NOT_FOUND;
-import static com.hcmute.drink.constant.ErrorConstant.PRODUCT_NOT_FOUND;
+import static com.hcmute.drink.constant.ErrorConstant.*;
 
 @Service
 @RequiredArgsConstructor
 public class ProductService implements IProductService {
     private final ProductRepository productRepository;
-    private final CloudinaryUtils cloudinaryUtils;
+    private final CloudinaryService cloudinaryService;
     private final CategoryService categoryService;
-    private final ModelMapperUtils modelMapperUtils;
+    private final ModelMapperService modelMapperService;
     private final SequenceService sequenceService;
     private final ProductSearchService productSearchService;
     @Autowired
@@ -45,14 +49,18 @@ public class ProductService implements IProductService {
 
 
     public ProductCollection getById(String id) {
-        return productRepository.findById(id)
+        return productRepository.getById(id)
                 .orElseThrow(() -> new CustomException(ErrorConstant.NOT_FOUND + id));
+    }
+
+    public ProductCollection saveProduct(ProductCollection data) {
+        return productRepository.save(data);
     }
 
     public static long getMinPrice(List<SizeEmbedded> sizeList) {
         long min = sizeList.get(0).getPrice();
         for (SizeEmbedded item : sizeList) {
-            if(min > item.getPrice()) {
+            if (min > item.getPrice()) {
                 min = item.getPrice();
             }
         }
@@ -65,9 +73,8 @@ public class ProductService implements IProductService {
     @Override
     public void createProduct(CreateProductRequest body, MultipartFile image) {
 
-        ProductCollection data = modelMapperUtils.mapClass(body, ProductCollection.class);
+        ProductCollection data = modelMapperService.mapClass(body, ProductCollection.class);
         Date currentDate = new Date();
-        long currentTimeMillis = currentDate.getTime();
 
         byte[] originalImage = new byte[0];
         try {
@@ -75,19 +82,23 @@ public class ProductService implements IProductService {
             byte[] newImage = ImageUtils.resizeImage(originalImage, 200, 200);
             byte[] newThumbnail = ImageUtils.resizeImage(originalImage, 200, 200);
 
-            categoryService.getById(data.getCategoryId().toString());
+            CategoryCollection categoryCollection = categoryService.getById(data.getCategoryId().toString());
+            if (categoryCollection.isCanDelete()) {
+                categoryCollection.setCanDelete(false);
+                categoryService.saveCategory(categoryCollection);
+            }
 
-            HashMap<String, String> imageUploaded = cloudinaryUtils.uploadFileToFolder(
+            HashMap<String, String> imageUploaded = cloudinaryService.uploadFileToFolder(
                     CloudinaryConstant.PRODUCT_PATH,
-                    data.getName() + "_" + currentTimeMillis,
+                    StringUtils.generateFileName(body.getName(), "product"),
                     newImage
             );
             ImageEmbedded imageEmbedded = new ImageEmbedded(imageUploaded.get(CloudinaryConstant.PUBLIC_ID), imageUploaded.get(CloudinaryConstant.URL_PROPERTY));
             data.setImage(imageEmbedded);
 
-            HashMap<String, String> thumbUploaded = cloudinaryUtils.uploadFileToFolder(
+            HashMap<String, String> thumbUploaded = cloudinaryService.uploadFileToFolder(
                     CloudinaryConstant.PRODUCT_PATH,
-                    data.getName() + "_thumb_" + currentTimeMillis,
+                    StringUtils.generateFileName(body.getName(), "product_thumb"),
                     newThumbnail
             );
 
@@ -135,57 +146,76 @@ public class ProductService implements IProductService {
     }
 
     @Override
-    public List<GetAllProductsResponse> getAllProducts(int page, int size) {
+    public GetProductListResponse getProductList(String key, int page, int size, String categoryId, ProductStatus productStatus) {
         int skip = (page - 1) * size;
         int limit = size;
-        return productRepository.getAllProducts(skip, limit);
+        String categoryIdRegex = RegexUtils.generateFilterRegexString(categoryId != null ? categoryId : "");
+        String productStatusRegex = RegexUtils.generateFilterRegexString(productStatus != null ? productStatus.toString() : "");
+        if (key == null) {
+            return productRepository.getProductList(skip, limit, categoryIdRegex, productStatusRegex);
+        } else {
+            return searchProduct(key, page, size, categoryIdRegex, productStatusRegex);
+        }
     }
 
     @Transactional
     @Override
     public void deleteProductById(String id) {
         ProductCollection product = getById(id);
-        ImageEmbedded image = product.getImage();
-
-        productRepository.deleteById(id);
-        productSearchService.deleteProduct(id);
-
-        try {
-            cloudinaryUtils.deleteImage(image.getId());
-        } catch (IOException e) {
-            productSearchService.createProduct(product);
-            throw new RuntimeException(e);
+        if (product.isCanDelete()) {
+            product.setDeleted(true);
+            productRepository.save(product);
+            productSearchService.deleteProduct(id);
+        } else {
+            throw new CustomException(CANT_DELETE);
         }
+    }
+
+    // TODO: kiểm tra lại trường hợp có 1 id không thể xóa
+    @Override
+    public DeleteSomeProductResponse deleteSomeProductById(List<String> productIdList) {
+        int successCount = 0;
+        for (String id : productIdList) {
+            try {
+                deleteProductById(id);
+                successCount++;
+            } catch (Exception e) {
+
+            }
+        }
+        return DeleteSomeProductResponse.builder().successCount(successCount).build();
     }
 
     @Transactional
     @Override
-    public void updateProductById(UpdateProductRequest data, String id) {
-        ProductCollection product = productRepository.findById(id).orElse(null);
+    public void updateProductById(UpdateProductRequest body, String id) {
+        ProductCollection product = productRepository.getById(id).orElse(null);
 
         if (product == null) {
             throw new CustomException(NOT_FOUND);
         }
 
-        modelMapperUtils.map(data, product);
+        modelMapperService.map(body, product);
 
-        if (data.getImage() != null) {
+        if (body.getImage() != null) {
             ImageEmbedded oldImage = product.getImage();
             ImageEmbedded oldThumbnail = product.getThumbnail();
 
             try {
-                cloudinaryUtils.deleteImage(oldImage.getId());
-                cloudinaryUtils.deleteImage(oldThumbnail.getId());
+                cloudinaryService.deleteImage(oldImage.getId());
+                cloudinaryService.deleteImage(oldThumbnail.getId());
 
-                byte[] originalImage = data.getImage().getBytes();
+                byte[] originalImage = body.getImage().getBytes();
 
                 byte[] newImage = ImageUtils.resizeImage(originalImage, 200, 200);
-                HashMap<String, String> fileUploaded = cloudinaryUtils.uploadFileToFolder(CloudinaryConstant.PRODUCT_PATH, data.getName(), newImage);
+                HashMap<String, String> fileUploaded = cloudinaryService.uploadFileToFolder(CloudinaryConstant.PRODUCT_PATH,
+                        StringUtils.generateFileName(body.getName(), "product"), newImage);
                 ImageEmbedded imageEmbedded = new ImageEmbedded(fileUploaded.get(CloudinaryConstant.PUBLIC_ID), fileUploaded.get(CloudinaryConstant.URL_PROPERTY));
                 product.setImage(imageEmbedded);
 
                 byte[] thumbnailImage = ImageUtils.resizeImage(originalImage, 200, 200);
-                HashMap<String, String> thumbnailUploaded = cloudinaryUtils.uploadFileToFolder(CloudinaryConstant.PRODUCT_PATH, data.getName(), thumbnailImage);
+                HashMap<String, String> thumbnailUploaded = cloudinaryService.uploadFileToFolder(CloudinaryConstant.PRODUCT_PATH,
+                        StringUtils.generateFileName(body.getName(), "product"), thumbnailImage);
                 ImageEmbedded thumbnailEmbedded = new ImageEmbedded(thumbnailUploaded.get(CloudinaryConstant.PUBLIC_ID), thumbnailUploaded.get(CloudinaryConstant.URL_PROPERTY));
                 product.setThumbnail(thumbnailEmbedded);
             } catch (IOException e) {
@@ -201,14 +231,20 @@ public class ProductService implements IProductService {
 
     @Override
     public List<GetAllVisibleProductResponse> searchProductVisible(String key, int page, int size) {
-        return modelMapperUtils.mapList(productSearchService.searchVisibleProduct(key, page, size), GetAllVisibleProductResponse.class);
+        return modelMapperService.mapList(productSearchService.searchVisibleProduct(key, page, size), GetAllVisibleProductResponse.class);
     }
+
     @Override
-    public List<GetAllProductsResponse> searchProduct(String key, int page, int size) {
-        return modelMapperUtils.mapList(productSearchService.searchProduct(key, page, size), GetAllProductsResponse.class);
+    public GetProductListResponse searchProduct(String key, int page, int size, String categoryIdRegex, String productStatusRegex) {
+        Page<ProductIndex> productPage = productSearchService.searchProduct(key, categoryIdRegex, productStatusRegex, page, size);
+        GetProductListResponse resultPage = new GetProductListResponse();
+        resultPage.setTotalPage(productPage.getTotalPages());
+        resultPage.setProductList(modelMapperService.mapList(productPage.getContent(), GetProductListResponse.Product.class));
+        return resultPage;
     }
+
     @Override
-    public List<GetAllVisibleProductResponse> getTopProductQuantityOrder(int quantity)  {
+    public List<GetAllVisibleProductResponse> getTopProductQuantityOrder(int quantity) {
         List<GetAllVisibleProductResponse> resData = orderService.getTopProductQuantityOrder(quantity);
         if (resData.size() < quantity) {
             Iterator itr = resData.iterator();
